requirements.txt
alembic==1.12.0
asgiref==3.7.2
async-timeout==4.0.3
coverage==7.3.1
Django==4.2.6
django-filter==23.3
django-polymorphic==3.1.0
django-redis==5.4.0
drf-yasg==1.21.7
django-rest-framework==0.1.0    
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
exceptiongroup==1.1.3
greenlet==3.0.0
inflection==0.5.1
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
mysqlclient==2.1.1
packaging==23.2
pluggy==1.3.0
protobuf==4.21.12
pytest==7.4.2
pytz==2023.3.post1
PyYAML==6.0.1
python-dotenv==1.0.0
redis==5.0.1
SQLAlchemy==2.0.22
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
